-- renamed some modules to the right name,
